Reduce boundary churn in order creation and payment wiring
OrdersService mixed transaction workflow, post-commit notifications,
and read-side queries, while PaymentsModule held both checkout policy
and adapter selection. This splits those responsibilities into
small collaborators so future order or gateway changes land in one
place without widening the regression surface.

Constraint: Keep runtime checkout and order-creation behavior unchanged during the refactor
Rejected: Bigger domain-service split across order reads/writes | too broad for one issue branch
Confidence: medium
Scope-risk: moderate
Reversibility: clean
Directive: Keep checkout visibility policy out of PaymentsModule and keep post-commit effects outside the transactional workflow
Tested: cd backend && npm run build && npm run test -- --runInBand src/modules/orders/__tests__/orders.service.spec.ts src/modules/payments/__tests__/payments.module.spec.ts src/modules/payments/__tests__/payment-gateway.provider.spec.ts
Tested: cd backend && npm run lint -- src/modules/orders src/modules/payments
Not-tested: Full backend test suite and e2e payment flows
