+++ backend/src/modules/payments/__tests__/payment-gateway.provider.spec.ts
@@ -1,0 +1,36 @@
+import type { PaymentConfig } from '../../../config/payment.config';
+import { selectPaymentGatewayAdapter } from '../payment-gateway.provider';
+
+function createConfig(gateway: PaymentConfig['gateway']): PaymentConfig {
+  return { gateway } as PaymentConfig;
+}
+
+describe('selectPaymentGatewayAdapter', () => {
+  const adapters = {
+    mock: { name: 'mock' },
+    toss: { name: 'toss' },
+    stripe: { name: 'stripe' },
+    inicis: { name: 'inicis' },
+    naverpay: { name: 'naverpay' },
+    paypal: { name: 'paypal' },
+    eximbay: { name: 'eximbay' },
+  };
+
+  it.each([
+    ['mock', adapters.mock],
+    ['toss', adapters.toss],
+    ['stripe', adapters.stripe],
+    ['inicis', adapters.inicis],
+    ['naverpay', adapters.naverpay],
+    ['paypal', adapters.paypal],
+    ['eximbay', adapters.eximbay],
+  ] as const)('returns %s adapter', (gateway, expected) => {
+    expect(selectPaymentGatewayAdapter(createConfig(gateway), adapters as never)).toBe(expected);
+  });
+
+  it('throws for unknown gateway values', () => {
+    expect(() =>
+      selectPaymentGatewayAdapter({ gateway: 'unknown' } as unknown as PaymentConfig, adapters as never),
+    ).toThrow('Unknown PAYMENT_GATEWAY: unknown');
+  });
+});
